qlink_chatbot/utils/cloudflare_client.py
import boto3
from dotenv import load_dotenv
import os

load_dotenv()

s3 = boto3.client(
    "s3",
    endpoint_url="https://b166538fe6434f1c027568fda8861597.r2.cloudflarestorage.com/jr-chatbot",
    aws_access_key_id=os.environ.get("R2_ACCESS_KEY"),
    aws_secret_access_key=os.environ.get("R2_SECRET_KEY"),
    region_name="auto"
)
