requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.118.2
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1

openai
pinecone
pymongo==4.10.1
httpx==0.27.0
asyncio
python-docx
python-multipart

boto3